requirements.txt
numpy
tqdm
torch
scikit-learn==1.2.2
matplotlib
xgboost
botorch
wandb
more-itertools==10.2.0
pygame==2.5.2
Box2D==2.3.10
gym==0.26.2
flax
chex
google-vizier
equinox
cvxpy
hydra
hydra-core
omegaconf
mlflow